Make Quick Office component extensions on Chrome OS
BUG=179070
TEST=manual

Review URL: https://codereview.chromium.org/12611007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@188960 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d6fa9528f3fe71966e1f180aeca1f55134c73f39

+++ chrome/test/functional/doc_viewing.py
@@ -22,9 +22,11 @@ class DocViewingTest(pyauto.PyUITest):
     pyauto.PyUITest.setUp(self)
     extension_path = '/opt/google/chrome/extensions'\
         '/gbkeegbaiigmenfmjfclcdgdpimamgkj.crx'
-    ext_id = self.InstallExtension(extension_path, from_webstore=True)
-    self.assertTrue(ext_id, msg='Failed to install extension %s' %
-                    extension_path)
+    # If crx file with doesn't exist, component extensions should be used.
+    if os.path.exists(extension_path):
+      ext_id = self.InstallExtension(extension_path, from_webstore=True)
+      self.assertTrue(ext_id, msg='Failed to install extension %s' %
+                      extension_path)
 
   def _GetFullPageFileBrowser(self):
     """Display the full page file browser.
